Move key.go from utils to internal/testutils
