+++ internal/testutils/key.go
@@ -1,7 +1,7 @@
-// (c) 2024, Ava Labs, Inc. All rights reserved.
+// (c) 2024-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package utils
+package testutils
 
 import (
 	"crypto/ecdsa"
