+++ sync/syncutils/test_trie.go
@@ -13,7 +13,7 @@ import (
 	"github.com/ava-labs/libevm/trie"
 	"github.com/ava-labs/libevm/trie/trienode"
 	"github.com/ava-labs/libevm/triedb"
-	"github.com/ava-labs/subnet-evm/utils"
+	"github.com/ava-labs/subnet-evm/internal/testutils"
 	"github.com/holiman/uint256"
 
 	"github.com/ava-labs/libevm/common"
@@ -146,12 +146,12 @@ func CorruptTrie(t *testing.T, diskdb ethdb.Batcher, tr *trie.Trie, n int) {
 func FillAccounts(
 	t *testing.T, trieDB *triedb.Database, root common.Hash, numAccounts int,
 	onAccount func(*testing.T, int, types.StateAccount) types.StateAccount,
-) (common.Hash, map[*utils.Key]*types.StateAccount) {
+) (common.Hash, map[*testutils.Key]*types.StateAccount) {
 	var (
 		minBalance  = uint256.NewInt(3000000000000000000)
 		randBalance = uint256.NewInt(1000000000000000000)
 		maxNonce    = 10
-		accounts    = make(map[*utils.Key]*types.StateAccount, numAccounts)
+		accounts    = make(map[*testutils.Key]*types.StateAccount, numAccounts)
 	)
 
 	tr, err := trie.NewStateTrie(trie.TrieID(root), trieDB)
@@ -175,7 +175,7 @@ func FillAccounts(
 			t.Fatalf("failed to rlp encode account: %v", err)
 		}
 
-		key := utils.NewKey(t)
+		key := testutils.NewKey(t)
 		tr.MustUpdate(key.Address[:], accBytes)
 		accounts[key] = &acc
 	}
